move helper functions into shared extension
